Rename router
