+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { RouterProvider } from 'react-router-dom';
 
-import router from './AppRouter';
+import router from './router';
 import Header from './components/layout/Header';
 
 function App() {
